Cleanly separate types between server-side only and client-side
